Add /tags page showing all avaliable tags
This is another page that is only used to make the URL structure complete,
just like /categories.
These pages only show the available tags/categories and link to the respective
index pages.

These pages are not meant to be linked to directly but serve only to make the
URL structure complete. Otherwise a URL like /tags/first-tag would have a
404 page in the URL structure.