Fix createPages to really return a promise
I actually am not sure why it worked before. But once I removed the async statement it
was not working any more. Also, apparently createPages needs to return a promise to
work properly. I looked into how `createWagtailPages` is defined and used that approach.
Basically it is not just chaining actions based on the promise that is returned by the
`graphql()` function.