Remove streamfield form blog page to prevent error
This error was dicovered in the simple page, but it was another one
than the one I am having here. The error here is still not resolved.

I am going to build the simple frontend step by step toward the one
here (not that this is more complicated). I hope to figure out what
causes the error I am having here.