Add image work around back in to fix preview
This is not an issue with the StreamField per se, but an issue with
having two images in the same query. When this is the case, then
the request to the backend is malformed.

https://github.com/GrappleGQL/gatsby-source-wagtail/issues/19#issuecomment-663930385

I need a workaround until this works right with the preview.

Using is in the stream field handling just felt natural as I had that
workaround before. This time, this is not an issue specifically with
the stream field. This issue always occurs on the second `imageFile`
block that is queried.

Also, this issue is limited to the preview. The requests with two
`imageFile` fields work fine for normal page rendering. This issue
is probably there because in preview no actual image processing can
happen.