Use slugify to create heading id
For some reason, there was an issue in Chrome when using `replaceAll`.
Apparently that is not really a standard JS string method. This only
exists in Firefox.

With slugify it works on Chrome too.

Also, there is probably a bit more functionality in this then only
replacing a space with -. I have already notices that is also removed
slashes.