Add some kind of handling for the struct block
Because the blocks field can not be queried, I need to deal with the
JSON rawValue as it is coming from the backend. This also means that
the image is not identified with a good value (like the ID) and I
would need to create a special solution for this stupid case right now
... and I really do not feel like that. I think my time is going to be
invested better elsewhere.

Also, this kind of issue can probablu be solved by choosing a different
type of block on the backend.

It would also not be an issue to create a solution to it. It is just a
little tedious right now for a project that does not matter. Something
that I did learn though is that the JSON in the rawValue also does still
have some formatting issue (single quotes instead of double) that need
to be converted before the JSON can be parsed.