Create Image component that abstracts the querying of image data fro the pages
I thought this would maybe more stable than having thiese queries in the page
data, but it turns out this is not the case. The error is still mysterious.
While the gallery image can be queried in the GraphQL IDEs without issue,
the array is empty in when this is included in the page...