Inlined the hero image extraction
For some reason not the gallery image array is now not empty...
But I have not even modified it...

I am not sure why, but I have a feeling that this error had something todo with the use of `shift` to retrieve the first image. I guess if the query response is cached somewhere, then using shift will empty the returned array over time making this break after several runs.
So now I am accessing the first item by index. 🤞 that it will hold...