Moving the image query back into the page query
Since this was not the cause of the error, it makes no sense to limit the
querying in that way. This also allows to define other query features
(like using the ImageSharp feature to grayscale the image) in the page query
without having to define a separate component for each of these.