Download all documents from Wagtail to static dir
The files in the static dir are copied directly into the public dir during build.
This makes the files in that folder available with the path (relative to the
static dir) as their URL.

This means I am serving all documents directly with the frontend. I do not need to
serve them with the backend, which would not benefit from the same CDN effects
as the frontend.

This should make the site a lot more effective.